Implementing Exploratory Player with No End Node
Algorithm is:
Archimedian spiral based traversal of the mapto increase known space on the map and dash to finish line once end is know.
